Split parsed elements into modules
